Segundo Commit: Módulo de Compras

+++ views/permissions.php
@@ -5,23 +5,47 @@
     <div class="tabitem">Permissões</div>
 </div>
 <div class="tabcontent">
-    <div class="tabbody style=display:block;">
-        GRUPOS DE PERMISSÕES
-    </div>
-    <div class="tabbody">
-
-        <a href="<?= BASE_URL; ?>/permissions/add">Adicionar Permissão</a>
+    <div class="tabbody" style=display:block;>
+        <div class="button"><a href="<?= BASE_URL; ?>/permissions/add_group" >
+        <span class="icon-add"> Adicionar Grupo de Permissões</span></a></div>
 
         <table border="0" width="100%">
             <tr>
-                <th>Nome da Permissão</th>
+                <th>Grupo de Permissões</th>
+                <th>Ações</th>
+            </tr>
+            <?php foreach($permissions_groups_list as $item): ?>
+            <tr>
+                <td><?= $item['name']; ?></td>
+                <td style="width: 225px; text-align: center;">
+                    <div class="button button_small">
+                        <a href="<?= BASE_URL; ?>/permissions/edit_group/<?= $item['id']; ?>"><span class="icon-edit"> Editar</span></a>
+                    </div>
+                    <div class="button button_small">
+                        <a href="<?= BASE_URL; ?>/permissions/delete_group/<?= $item['id']; ?>" onclick="return confirm('Tem certeza que deseja excluir?')"><span class="icon-del"> Excluir</span></a>
+                    </div>
+                </td>
+            </tr>
+            <?php endforeach; ?>
+
+        </table>
+    </div>
+    <div class="tabbody">
+
+        <div class="button"><a href="<?= BASE_URL; ?>/permissions/add" ><span class="icon-add"> Adicionar Permissão</span></a></div>
+
+        <table border="0" width="100%">
+            <tr>
+                <th>Permissão</th>
                 <th>Ações</th>
             </tr>
             <?php foreach($permissions_list as $item): ?>
             <tr>
                 <td><?= $item['name']; ?></td>
-                <td>
-                    <a href="<?= BASE_URL; ?>/permissions/delete/<?= $item['id']; ?>">[excluir]</a>
+                <td style="width: 225px; text-align: center;">
+                    <div class="button button_small">
+                        <a href="<?= BASE_URL; ?>/permissions/delete/<?= $item['id']; ?>" onclick="return confirm('Tem certeza que deseja excluir?')"><span class="icon-del"> Excluir</span></a>
+                    </div>
                 </td>
             </tr>
             <?php endforeach; ?>
